+++ src/puuid/__init__.py
@@ -4,12 +4,12 @@ pUUID - Prefixed UUIDs for Python with Pydantic & SQLAlchemy support.
 Author: Jendrik Potyka, Fabian Preiss
 """
 
-__version__ = "1.0.0"
+__version__ = "1.1.0"
 __author__ = "Jendrik Potyka, Fabian Preiss"
 
 
 from puuid.base import (
-    PUUID,
+    PUUIDBase,
     PUUIDError,
     PUUIDv1,
     PUUIDv3,
@@ -20,8 +20,11 @@ from puuid.base import (
     PUUIDv8,
 )
 
+PUUID = PUUIDBase  # backwards compatibility
+
 __all__ = [
     "PUUID",
+    "PUUIDBase",
     "PUUIDv1",
     "PUUIDv3",
     "PUUIDv4",