+++ src/puuid/sqlalchemy.py
@@ -3,14 +3,14 @@ from typing import final, override
 from sqlalchemy.engine.interfaces import Dialect
 from sqlalchemy.types import String, TypeDecorator
 
-from puuid.base import PUUID
+from puuid.base import PUUIDBase
 
 _SEPARATOR_LENGTH = 1
 _UUID_LENGTH = 36
 
 
 @final
-class SqlPUUID(TypeDecorator[PUUID[str]]):
+class SqlPUUID[TPrefix: str](TypeDecorator[PUUIDBase[TPrefix]]):
     """
     SQLAlchemy type for storing Prefixed UUIDs.
 
@@ -21,16 +21,19 @@ class SqlPUUID(TypeDecorator[PUUID[str]]):
     impl = String
     cache_ok = True
 
-    puuid_cls: type[PUUID[str]]
+    puuid_cls: type[PUUIDBase[TPrefix]]
 
-    def __init__(self, puuid_cls: type[PUUID[str]], prefix_length: int = 4) -> None:
+    def __init__(
+        self, puuid_cls: type[PUUIDBase[TPrefix]], prefix_length: int = 4
+    ) -> None:
         """
         Initialize the SqlPUUID type.
 
         Parameters
         ----------
-        puuid_cls : type[PUUID[str]]
-            The pUUID class (e.g., `UserUUID`) to associate with this column.
+        puuid_cls : type[PUUIDBase[TPrefix]]
+            The pUUID class (e.g., `PUUIDv4[Literal["user"]]`) to associate with this
+            column.
         prefix_length : int, default 4
             The length of the prefix string to calculate the column width.
         """
@@ -40,7 +43,7 @@ class SqlPUUID(TypeDecorator[PUUID[str]]):
 
     @override
     def process_bind_param(
-        self, value: PUUID[str] | None, dialect: Dialect
+        self, value: PUUIDBase[TPrefix] | None, dialect: Dialect
     ) -> str | None:
         if value is None:
             return None
@@ -49,7 +52,7 @@ class SqlPUUID(TypeDecorator[PUUID[str]]):
     @override
     def process_result_value(
         self, value: str | None, dialect: Dialect
-    ) -> PUUID[str] | None:
+    ) -> PUUIDBase[TPrefix] | None:
         if value is None:
             return None
         return self.puuid_cls.from_string(value)